Extract CodeHighlighting
