+++ Sources/SlideUI/LegacyCode.swift
@@ -115,39 +115,3 @@ private struct ResolvedLegacyCodeStyle<Style: LegacyCodeStyle>: View {
         style.makeBody(configuration: configuration)
     }
 }
-
-// MARK: - CodeHighlighting
-
-extension View {
-
-    public func codeHighlighting(_ style: some CodeHighlighting) -> some View {
-        environment(\.codeHighlighting, style)
-    }
-}
-
-extension Scene {
-
-    public func codeHighlighting(_ style: some CodeHighlighting) -> some Scene {
-        environment(\.codeHighlighting, style)
-    }
-}
-
-public protocol CodeHighlighting: DynamicProperty {
-    func color(for token: Token) -> Color
-}
-
-private struct CodeHighlightingKey: EnvironmentKey {
-    static var defaultValue: any CodeHighlighting = DefaultCodeHighlighting()
-}
-
-private struct DefaultCodeHighlighting: CodeHighlighting {
-    func color(for token: Token) -> Color { .black }
-}
-
-extension EnvironmentValues {
-
-    fileprivate var codeHighlighting: any CodeHighlighting {
-        get { self[CodeHighlightingKey.self] }
-        set { self[CodeHighlightingKey.self] = newValue }
-    }
-}
